// circabc-frontend/src/app/core/generated/circabc/model/autoUploadConfiguration.ts
/**
 * CIRCABC REST API
 * This is the first version of the CIRCABC REST API used by the new User Interface 
 *
 * The version of the OpenAPI document: 4.2.0
 * Contact: DIGIT-CIRCABC-SUPPORT@ec.europa.eu
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AutoUploadConfiguration { 
    idConfiguration?: number;
    status?: number;
    igName?: string;
    fileId?: string;
    parentId?: string;
    fileName?: string;
    ftpHost?: string;
    ftpPort?: number;
    ftpUsername?: string;
    ftpPassword?: string;
    ftpPath?: string;
    dateRestriction?: string;
    dayChoice?: number;
    hourChoice?: number;
    autoExtract?: boolean;
    jobNotifications?: boolean;
    emails?: string;
}